pyproject.toml
[project]
name = "xai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.128.0",
    "langchain>=1.2.8",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "tweepy>=4.14.0",
]
